move and improve examples
