+++ RustPlusApi/Examples/GetEntityInfo/GetEntityInfo.csproj
