+++ RustPlusApi/Examples/GetInfo/GetInfo.csproj
