+++ RustPlusApi/Examples/GetTeamInfo/GetTeamInfo.csproj
