+++ RustPlusApi/Examples/SendTeamChat/SendTeamChat.csproj
@@ -1,9 +1,15 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
+    <OutputType>Exe</OutputType>
     <TargetFramework>net8.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
   </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\RustPlusApi\RustPlusApi.csproj" />
+    <ProjectReference Include="..\__Constants\__Constants.csproj" />
+  </ItemGroup>
 
 </Project>
