--- RustPlusApi/Exemples/__Constants/RustPlusConst.cs
@@ -1,10 +1,0 @@
-﻿namespace __Constants
-{
-    public record RustPlusConst
-    {
-        public const string Ip = "";
-        public const int Port = 0;
-        public const ulong PlayerId = 0;
-        public const int PlayerToken = 0;
-    }
-}
